core: use more standard items
